Implement normal map detection based on image data.
The current 'use_normal' flag is very inexact, resulting in both
normal maps being used as color data (quite obvious), and color
textures being used as normal maps - the latter is worse as it's
not immediately apparent.

Since it's not obvious how to do this properly, instead scan
the image data to determine if it 'looks like' a normal map.
After some tweaking of thresholds this seems to work quite well.