Fix object placement for valcA08bX_h.mlx and bone handling.
Object parenting must always use the accum_matrix of the
bone, even if the bone is somehow missing from the armature.
That in turn happens if the script tries to create a zero-
length bone.

Change parenting code to be robust in all situations, and
redo the bone tail computations to avoid zero length bone
issues, and overall try to place bones more sensibly, esp.
in environment maps.