Support splitting materials into copies with and without vertex colors.
Sometimes the same material is used by multiple meshes, some of which
have colors, and some had them pruned due to lack of useful data. This
causes red meshes to appear when a material expects a color, but there
is none. To avoid this, create two materials in such cases.

Example: evmap09_02.mxe