Basic update of the importer for blender 2.83.
The file reading part worked as is, but parts were slightly
refactored to group values into vector tuples, making the
importer code more sensible. The material code had to be
rewritten due to the switch to pure nodes.

The texconv.exe binary is now included into the addon for
convenience, but this and other VC4 code is not tested.